MTDSA-4024 - move to renamed packages
