+++ app/v1/models/response/getIncomeTaxCalc/CalculationSummary.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package v1.models.backend.selfAssessment
+package v1.models.response.getIncomeTaxCalc
 
 import play.api.libs.json.{Json, OFormat}
